requirements.txt
flask
requests
langchain
langchain_community
openai
python-dotenv